# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Inventory {
  id: number;
  name: string;
  location: string;
  date: string;
  createdAt: string;
}

export interface InventoryItem {
  id: number;
  inventoryId: number;
  sku: string;
  descripcion: string;
  categoria: string;
  cantidadTeorica: number;
  cantidadFisica: number;
}

export interface InventoryWithItems {
  id: number;
  name: string;
  location: string;
  date: string;
  createdAt: string;
  items: InventoryItem[];
}

export interface InventoryUpload {
  name: string;
  location: string;
  date: string;
}

export interface ItemUpdate {
  /** @minimum 0 */
  cantidadFisica: number;
}