# lib/api-zod/src/generated/types/inventory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inventory {
  id: number;
  name: string;
  location: string;
  date: string;
  createdAt: string;
}