# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItem {
  id: number;
  inventoryId: number;
  sku: string;
  descripcion: string;
  categoria: string;
  cantidadTeorica: number;
  cantidadFisica: number;
}