# lib/api-zod/src/generated/types/inventoryWithItems.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItem } from './inventoryItem';

export interface InventoryWithItems {
  id: number;
  name: string;
  location: string;
  date: string;
  createdAt: string;
  items: InventoryItem[];
}