# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ItemUpdate {
  /** @minimum 0 */
  cantidadFisica: number;
}